requirements.txt
pytest
simple-salesforce
pyyaml